New Architecture for drawer and top bar in BaseActivity
