--- frontend-TreasureHunt/app/src/main/java/com/geoschnitzel/treasurehunt/base/ToolBarFragment.java
@@ -1,127 +1,0 @@
-package com.geoschnitzel.treasurehunt.base;
-
-import android.app.Activity;
-import android.content.Context;
-import android.net.Uri;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.geoschnitzel.treasurehunt.IBasePresenter;
-import com.geoschnitzel.treasurehunt.IBaseView;
-import com.geoschnitzel.treasurehunt.R;
-import static com.google.common.base.Preconditions.checkNotNull;
-
-/**
- * A simple {@link Fragment} subclass.
- * Activities that contain this fragment must implement the
- * {@link ToolBarFragment.OnFragmentInteractionListener} interface
- * to handle interaction events.
- * Use the {@link ToolBarFragment#newInstance} factory method to
- * create an instance of this fragment.
- */
-public class ToolBarFragment extends Fragment implements BaseContract.View {
-
-    private IBasePresenter mPresenter;
-
-    private static final String ARG_PARAM1 = "schnitzelCount";
-
-    private int mSchnitzelCount;
-
-    private OnFragmentInteractionListener mListener;
-
-    public ToolBarFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public void setPresenter(BaseContract.Presenter presenter) {
-        mPresenter = checkNotNull(presenter);
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param schnitzelCount Current amount of schnitzel for a user
-     * @return A new instance of fragment ToolBarFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static ToolBarFragment newInstance(Integer schnitzelCount) {
-        ToolBarFragment fragment = new ToolBarFragment();
-        Bundle args = new Bundle();
-        args.putInt(ARG_PARAM1, schnitzelCount);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mSchnitzelCount = getArguments().getInt(ARG_PARAM1);
-        }
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View root = inflater.inflate(R.layout.fragment_tool_bar, container, false);
-
-        Toolbar toolbar = root.findViewById(R.id.toolbar);
-        ((AppCompatActivity)getActivity()).setSupportActionBar(toolbar);
-
-        ActionBar actionbar = ((AppCompatActivity)getActivity()).getSupportActionBar();
-        actionbar.setDisplayHomeAsUpEnabled(true);
-        actionbar.setHomeAsUpIndicator(R.drawable.ic_menu);
-        return root;
-    }
-
-    // TODO: Rename method, update argument and hook method into UI event
-    public void onButtonPressed(Uri uri) {
-        if (mListener != null) {
-            mListener.onFragmentInteraction(uri);
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        if (context instanceof OnFragmentInteractionListener) {
-            mListener = (OnFragmentInteractionListener) context;
-        } else {
-            throw new RuntimeException(context.toString()
-                    + " must implement OnFragmentInteractionListener");
-        }
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-
-
-
-    /**
-     * This interface must be implemented by activities that contain this
-     * fragment to allow an interaction in this fragment to be communicated
-     * to the activity and potentially other fragments contained in that
-     * activity.
-     * <p>
-     * See the Android Training lesson <a href=
-     * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
-     */
-    public interface OnFragmentInteractionListener {
-        // TODO: Update argument type and name
-        void onFragmentInteraction(Uri uri);
-    }
-}
