+++ frontend-treasurehunt/src/main/res/layout/activity_map.xml
@@ -13,8 +13,8 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <FrameLayout
-            android:id="@+id/frame_toolbar"
+        <include
+            layout="@layout/fragment_tool_bar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
 
@@ -23,8 +23,9 @@
             android:layout_height="wrap_content"/>
     </LinearLayout>
 
-    <!-- Navigation Drawer -->
-    <android.support.design.widget.NavigationView
+    <android.support.design.widget.NavigationView xmlns:tools="http://schemas.android.com/tools"
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
         android:id="@+id/nav_view"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
